apps/api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.31.0
sqlalchemy==2.0.35
python-multipart==0.0.12
pillow==11.0.0
pdfplumber==0.11.4
httpx==0.27.2
pandas==2.2.3
openpyxl==3.1.5
reportlab==4.2.5
rapidfuzz==3.10.0
python-dateutil==2.9.0
python-dotenv==1.0.1
